+++ skins/weewx-wdc/src/js/diagrams/types.ts
@@ -2,6 +2,14 @@ import type { Serie } from "@nivo/line";
 
 type precision = "day" | "week" | "month" | "year" | "alltime";
 
+type CalendarDiagramBaseProps = {
+  color: string[];
+  unit: string;
+  data: Array<{ day: string; value: number }>;
+  observation: string;
+  heading: string;
+};
+
 type DiagramBaseProps = {
   color: string[];
   unit?: string;
@@ -15,4 +23,4 @@ interface Series {
   y: number;
 }
 
-export { DiagramBaseProps, precision, Series };
+export { CalendarDiagramBaseProps, DiagramBaseProps, precision, Series };